pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "geoserver"
version = "0.0.1"
description = "Minimal python client to communicate with Geoserver."
authors = ["Arthur Dujardin <arthurdujardin.github@gmail.com>"]
maintainers = ["Arthur Dujardin <arthurdujardin.github@gmail.com>"]
homepage = "https://github.com/arthurdjn/geoserver-py/"
repository = "https://github.com/arthurdjn/geoserver-py/"
documentation = "https://github.com/arthurdjn/geoserver-py/"
license = "MIT"
readme = "README.md"
keywords = ["geoserver", "python", "client", "api"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
mypy = "^1.8.0"
ruff = "^0.1.14"
isort = "^5.13.2"
mkdocs = "^1.5.3"
mkdocs-material = "^9.5.6"
mkdocstrings-python = "^1.8.0"
types-requests = "^2.31.0.20240125"
ipykernel = "^6.29.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
seaborn = "^0.13.2"
mkdocs-jupyter = "^0.24.7"
mkdocs-simple-hooks = "^0.1.5"
mkdocs-exclude = "^1.0.2"
nbconvert = "^7.16.4"
dict2xml = "^1.7.5"
pre-commit = "^3.7.1"
bump-my-version = "^0.21.1"

[tool.black]
target-version = ["py310"]
line-length = 120

[tool.ruff]
target-version = "py310"
line-length = 120
ignore-init-module-imports = true

[tool.ruff.per-file-ignores]
"geoserver/**/__init__.py" = ["F401"]

[tool.ruff.pydocstyle]
convention = "google"

[tool.isort]
profile = "black"
line_length = 120
known_first_party = ["geoserver"]

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
no_implicit_optional = true
show_error_codes = true
ignore_missing_imports = true
explicit_package_bases = true
check_untyped_defs = true
implicit_reexport = true

[tool.pytest.ini_options]
addopts = ["--capture=no", "--color=yes", "-vv"]

[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
precision = 2
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == '__main__'",
]

[tool.bumpversion]
current_version = "1.0.1"
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = ["{major}.{minor}.{patch}"]
regex = false
ignore_missing_version = false
tag = true
sign_tags = false
tag_name = "{new_version}"
tag_message = "Bump version: {current_version} → {new_version}"
allow_dirty = true
commit = false
message = "Version updated from {current_version} to {new_version}"
commit_args = ""

[[tool.bumpversion.files]]
filename = "geoserver.__init__.py"
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'

[[tool.bumpversion.files]]
filename = "docker-compose.yml"
search = "descriptor:{current_version}"
replace = "descriptor:{new_version}"

